regroup _get_dimension*
