+++ tests/macrodata/test_pynsee_macrodata.py
@@ -16,8 +16,10 @@ from pynsee.macrodata._get_idbank_internal_data import (
     _get_idbank_internal_data,
 )
 from pynsee.macrodata._get_metadata import _get_dataset_metadata
-from pynsee.macrodata._get_dimension_vals import _get_dimension_values
-from pynsee.macrodata._get_dimensions import _get_dataset_dimension
+from pynsee.macrodata._get_dimensions import (
+    _get_dataset_dimension,
+    _get_dimension_values,
+)
 from pynsee.macrodata._download_idbanks import _download_idbank_list
 from pynsee.macrodata._get_dataset_list_internal import (
     _get_dataset_list_internal,
